+++ src/cli.ts
@@ -15,9 +15,10 @@ program
   .command("run", { isDefault: true })
   .description("Run the MX SSO API proxy server")
   .option("-p, --port <number>", "port number", process.env.PORT || "8089")
+  .option("--open-path <path>", "open browser after server starts", "")
   .option("--serve-local-files", "serve local files", false)
-  .action((options: { port: number; serveLocalFiles: boolean }) =>
-    run(options.port, options.serveLocalFiles),
+  .action((options: { port: number; openPath: string; serveLocalFiles: boolean }) =>
+    run(options.port, options.serveLocalFiles, options.openPath),
   )
 
 program.parse()